Move modal to modals folder
Additional fix to close modal on successful due date extension.
