+++ src/app/components/pages/quizzes/QuizTeacherFeedback.tsx
@@ -1,10 +1,6 @@
 import React, {useState} from "react";
 import {
-    closeActiveModal,
     getRTKQueryErrorMessage,
-    mutationSucceeded,
-    showSuccessToast,
-    useAppDispatch,
     useGetQuizAssignmentWithFeedbackQuery,
     useUpdateQuizAssignmentMutation
 } from "../../../state";
@@ -23,12 +19,11 @@ import {
     isQuestion,
     PATHS
 } from "../../../services";
+import {ExtendDueDateModal} from "../../elements/modals/ExtendDueDateModal";
 import {AppAssignmentProgress, AssignmentProgressPageSettingsContext, QuizFeedbackModes} from "../../../../IsaacAppTypes";
 import {teacherQuizzesCrumbs} from "../../elements/quiz/QuizAttemptComponent";
 import {formatDate} from "../../elements/DateString";
 import {Spacer} from "../../elements/Spacer";
-import {currentYear, DateInput} from "../../elements/inputs/DateInput";
-import range from "lodash/range";
 import {ResultsTable, passMark} from "../../elements/quiz/QuizProgressCommon";
 import {
     Alert,
@@ -39,9 +34,6 @@ import {
     DropdownMenu,
     DropdownToggle,
     Label,
-    Modal,
-    ModalBody,
-    ModalHeader,
     Row,
     UncontrolledButtonDropdown
 } from "reactstrap";
@@ -58,72 +50,6 @@ const feedbackNames: Record<QuizFeedbackMode, string> = {
     OVERALL_MARK: "Overall mark only",
     SECTION_MARKS: "Section-by-section mark breakdown",
     DETAILED_FEEDBACK: "Detailed feedback on each question",
-};
-
-type ExtendDueDateModalProps = {
-    isOpen: boolean;
-    toggle: () => void;
-    currDueDate: Date;
-    numericQuizAssignmentId: number;
-}
-const ExtendDueDateModal = (props: ExtendDueDateModalProps) => {
-    const {isOpen, toggle, currDueDate, numericQuizAssignmentId} = props;
-    const yearRange = range(currentYear, currentYear + 5);
-    const [dueDate, setDueDate] = useState<Date>(currDueDate);
-    const [updateQuiz, {isLoading: isUpdatingQuiz}] = useUpdateQuizAssignmentMutation();
-    const dispatch = useAppDispatch();
-
-    const setValidDueDate = () => {
-        if (isUpdatingQuiz || !numericQuizAssignmentId || !dueDate || currDueDate == dueDate) {
-            return;
-        }
-
-        if (currDueDate && (dueDate > currDueDate || dueDate > TODAY())) {
-            updateQuiz({quizAssignmentId: numericQuizAssignmentId, update: {dueDate: dueDate}})
-                .then((result) => {
-                    if (mutationSucceeded(result)) {
-                        dispatch(showSuccessToast(
-                            "Due date extended successfully", `This test is now due ${dueDate.toLocaleDateString()}.`
-                        ));
-                    }
-                    dispatch(closeActiveModal());
-                });
-        }
-    };
-
-    return <Modal isOpen={isOpen} toggle={toggle}>
-        <ModalHeader role={"heading"} className={"text-break"} close={
-            <button className={"text-nowrap close"} onClick={toggle}>
-                Close
-            </button>
-        }>
-            Extend due date?
-        </ModalHeader>
-        <ModalBody>
-            <p className="px-1">{`Are you sure you want to change the due date? This will extend the due date for all users this test is assigned to. ${dueDate <= TODAY()}`}</p>
-            <hr className="text-center"/>
-            <Container className="py-2">
-                <Label for="dueDate" className="pr-1">Extend the due date:
-                <DateInput id="dueDate" value={dueDate} invalid={dueDate && ((dueDate < currDueDate) || dueDate <= TODAY())}
-                    yearRange={yearRange} noClear disabled={isUpdatingQuiz} className="text-center"
-                    onChange={(e) => e.target.valueAsDate && setDueDate(e.target.valueAsDate)}
-                           />
-                </Label>
-                {dueDate && (dueDate < currDueDate || dueDate <= TODAY()) && 
-                <p className={"text-danger"}>
-                    You cannot set the due date to be earlier than the current due date or today.
-                </p>}
-                <Button
-                    className="mt-2 mb-2"
-                    block color="secondary"
-                    onClick={setValidDueDate}
-                    role={"button"}
-                    disabled={isUpdatingQuiz && dueDate && (dueDate < currDueDate)}>
-                    Extend due date
-                </Button>
-            </Container>
-        </ModalBody>
-    </Modal>;
 };
 
 export const QuizTeacherFeedback = ({user}: {user: RegisteredUserDTO}) => {
